Lift the state up, from FutureBookshelf to a Bookshelves component. Note: localStorage has stopped working
